[ASV-1570] Changed update size attribute type and migrated data base

+++ aptoide-database/src/main/java/cm/aptoide/pt/database/realm/Update.java
@@ -39,7 +39,7 @@ public class Update extends RealmObject {
   private long timestamp;
   private String md5;
   private String apkPath;
-  private double size;
+  private long size;
   private String updateVersionName;
   private int updateVersionCode;
   private boolean excluded;
@@ -63,7 +63,7 @@ public class Update extends RealmObject {
   }
 
   public Update(long id, String name, String icon, String packageName, String md5sum, String path,
-      double size, String versionName, String pathAlt, int versionCode, String rankName,
+      long size, String versionName, String pathAlt, int versionCode, String rankName,
       String mainObbFileName, String mainObbPath, String mainObbMd5, String patchObbFileName,
       String patchObbPath, String patchObbMd5, boolean appcUpgrade, boolean hasAppc) {
     this.appId = id;
@@ -151,11 +151,11 @@ public class Update extends RealmObject {
     this.apkPath = apkPath;
   }
 
-  public double getSize() {
+  public long getSize() {
     return size;
   }
 
-  public void setSize(double size) {
+  public void setSize(long size) {
     this.size = size;
   }
